renaming extensions and remove token credentials
