+++ src/ErniAcademy.Events.ServiceBus/ProcessorProvider/QueueProvider.cs
@@ -3,6 +3,9 @@ using ErniAcademy.Events.ServiceBus.ClientProvider;
 
 namespace ErniAcademy.Events.ServiceBus.ProcessorProvider;
 
+/// <summary>
+/// Provider to build a ServiceBusProcessor for ServiceBus queues
+/// </summary>
 public class QueueProvider : IServiceBusProcessorProvider
 {
     private readonly IServiceBusClientProvider _serviceBusClientProvider;
