+++ src/ErniAcademy.Events.StorageQueues/Extensions/ServiceCollectionExtensions.cs
@@ -13,7 +13,8 @@ namespace ErniAcademy.Events.StorageQueues.Extensions;
 public static class ServiceCollectionExtensions
 {
     /// <summary>
-    /// Extension method to configure IEventPublisher contract with StorageQueuePublisher by default will use connection string options to connect to Storage Queue
+    /// Extension method to configure IEventPublisher contract with StorageQueuePublisher by default will use connection string options 
+    /// to connect to Storage Queue, make sure ConnectionString of the StorageQueue is configure in the configuration section.
     /// </summary>
     /// <param name="services">the ServiceCollection</param>
     /// <param name="configuration">the Configuration used to bind and configure the options</param>
@@ -21,7 +22,7 @@ public static class ServiceCollectionExtensions
     /// <param name="sectionKey">the configuration section key to get the options</param>
     /// <param name="queueOptions">the QueueClient options for extra configuration</param>
     /// <returns>IServiceCollection</returns>
-    public static IServiceCollection AddEventsStorageQueues(this IServiceCollection services,
+    public static IServiceCollection AddEventsPublisherStorageQueues(this IServiceCollection services,
         IConfiguration configuration,
         ISerializer serializer,
         string sectionKey,
@@ -42,6 +43,17 @@ public static class ServiceCollectionExtensions
         return services;
     }
 
+    /// <summary>
+    /// Extension method to configure IEventSubscriber<TEvent> contract with StorageQueueSubscriber<TEvent> by default will use connection string options 
+    /// to connect to Storage Queue, make sure ConnectionString of the StorageQueue is configure in the configuration section.
+    /// </summary>
+    /// <typeparam name="TEvent">The generic event type to be subscribed</typeparam>
+    /// <param name="services">the ServiceCollection</param>
+    /// <param name="configuration">the Configuration used to bind and configure the options</param>
+    /// <param name="serializer">the serializer to be use</param>
+    /// <param name="sectionKey">the configuration section key to get the options</param>
+    /// <param name="queueOptions">the QueueClient options for extra configuration</param>
+    /// <returns>IServiceCollection</returns>
     public static IServiceCollection AddEventsSubscriberStorageQueues<TEvent>(this IServiceCollection services,
         IConfiguration configuration,
         ISerializer serializer,
