+++ tests/ErniAcademy.Events.IntegrationTests/RedisTests.cs
@@ -8,7 +8,7 @@ public class RedisTests : BaseTests
 {
     protected override IServiceCollection RegisterSut(IServiceCollection services, IConfiguration configuration)
     {
-        services.AddEventsRedis(configuration, _serializer, sectionKey: "Events:Redis");
+        services.AddEventsPublisherRedis(configuration, _serializer, sectionKey: "Events:Redis");
         services.AddEventsSubscriberRedis<DummyEvent>(configuration, _serializer, sectionKey: "Events:Redis");
         return services;
     }
